requirements.txt
# Core dependencies
requests>=2.28.0
click>=8.1.0
python-dotenv>=1.0.0
pyyaml>=6.0
rich>=13.0.0

# Optional extensions
tiktoken>=0.5.0  # For token counting with OpenAI models
pypdf>=3.15.0    # For PDF file support